% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run}
\alias{run}
\title{Run a power analysis simulation to show the probability of
detecting a somatic signature in a sample of mutations, when
the proportion of somatic mutations is known}
\usage{
run(proportion, mutations, signatures, outerloop = 100, stan.iter = 2000,
  stan.chains = 1, stan.warmup = 1000, credint = 0.9)
}
\arguments{
\item{proportion}{The proportion of somatic mutations in the simulation}

\item{mutations}{The total number of mutations (germline and somatic combined)}

\item{signatures}{A data frame or a matrix with the germline signature in
column 1 and the somatic signature in column 2}

\item{outerloop}{The number of outer loop iterations to run. Each iteration of the
outer loop simulates a new set of mutations with the given somatic proportion.}

\item{stan.iter}{The total number of samples generated by the Bayesian model 
per chain, per outer loop iteration. Warmup samples are discarded, so stan.iter
must be greater than stan.warmup}

\item{stan.chains}{The number of chains sampled by the Bayesian model}

\item{stan.warmup}{The number of samples discarded by each chain as warmup/burnin}

\item{credint}{The width of the credible interval used to generate statistics}
}
\description{
Run a power analysis simulation to show the probability of
detecting a somatic signature in a sample of mutations, when
the proportion of somatic mutations is known
}

